+++ test/twitter.test.ts
@@ -1,8 +1,6 @@
 import { test, expect } from "vitest"
 import { getFollowerCount } from "../src"
 
-process.env.FOLLOWER_COUNT_NO_SANDBOX = 'true'
-
 test("twitter", async () => {
   const count = await getFollowerCount({
     type: "twitter",