+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
-No unreleased changes.
+- fix(twitter): disable Chrome sandbox by default.
+- fix(twitter): wait for selector to appear.
 
 ## 0.1.0